requirements.txt
# Forge - Local AI Creative Suite
# Python 3.10+ required

# Core UI
nicegui>=2.0.0

# Native window (for packaged app)
pywebview>=5.0.0

# ComfyUI communication
websocket-client>=1.6.0
requests>=2.28.0

# Voice generation (Chatterbox TTS - optional, downloads on first use)
# torch>=2.0.0
# torchaudio>=2.0.0
# chatterbox-tts>=0.1.0

# Audio processing (for voice speed/pitch control)
# soundfile>=0.12.0
# Note: Also requires `brew install sound-touch` on macOS
